% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testInline.R
\name{testInline}
\alias{testInline}
\title{Write tests in your source}
\usage{
testInline(expr, file, dir = "testthat", eval = interactive())
}
\arguments{
\item{expr}{test code}

\item{file}{character with name of file in which expr will be stored}

\item{dir}{directory where you want your tests. Allways as sub directory of tests in your package root.}

\item{eval}{if you want to evaluate your testcode. This will call `devtools::test`}
}
\description{
This functions can be used in package development to keep your unit tests where your function definitions are.
}

